# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skillset API",
    "version": "0.1.0",
    "description": "A self-hosted Registry for Agent Skills. This is the contract the CLI depends on; the web interface uses the same routes and has no private data path of its own.\n\n**Conventions.** Every JSON key on the wire is snake_case, which means wire keys line up with Postgres column names. TypeScript types, interfaces, and classes are camelCase, so conversion happens once at the boundary in the shared module rather than ad hoc per call site.\n\n**Artifact bytes never pass through this API.** Publishing returns a presigned upload target that the caller writes to directly, and reading an Artifact redirects to a short-lived presigned location. The API therefore validates the metadata a caller claims about a Skill, never the Artifact itself (ADR-0001)."
  },
  "servers": [{ "url": "/api" }],
  "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "List Users",
        "description": "Requires at least the admin role.",
        "parameters": [{ "$ref": "#/components/parameters/Page" }],
        "responses": {
          "200": {
            "description": "A page of Users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items", "page", "page_size", "total"],
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
                    "page": { "type": "integer" },
                    "page_size": { "type": "integer" },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Create a User",
        "description": "Requires at least the admin role. The Registry generates the initial password and returns it once — there is no email delivery. The new User must change it before they can do anything else.\n\n`role` accepts `reader`, `writer`, or `admin` only — never `superadmin`, which exists on exactly one User and is set only by `/setup`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["first_name", "last_name", "email", "role"],
                "properties": {
                  "first_name": { "type": "string", "minLength": 1 },
                  "last_name": { "type": "string", "minLength": 1 },
                  "email": { "type": "string", "format": "email" },
                  "role": { "$ref": "#/components/schemas/AssignableRole" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created. `initial_password` is shown once and is not recoverable.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["user", "initial_password"],
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" },
                    "initial_password": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" },
          "409": {
            "description": "A User with that email already exists",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "tags": ["users"],
        "summary": "Read the authenticated User",
        "description": "Also the CLI's identity command: it distinguishes a wrong Registry location from a rejected Token. The role is resolved on this request, not read from the credential.",
        "responses": {
          "200": {
            "description": "The authenticated User",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      },
      "patch": {
        "tags": ["users"],
        "summary": "Correct your own names",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "first_name": { "type": "string", "minLength": 1 },
                  "last_name": { "type": "string", "minLength": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
    "/users/me/password": {
      "put": {
        "tags": ["users"],
        "summary": "Replace your own password",
        "description": "Clears `must_change_password`. Permitted while that flag is set — this is the route that resolves it.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["current_password", "new_password"],
                "properties": {
                  "current_password": { "type": "string" },
                  "new_password": { "type": "string", "minLength": 12 }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Password replaced" },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
    "/users/me/tokens": {
      "get": {
        "tags": ["users"],
        "summary": "List your own Tokens",
        "description": "Secrets are never included. `last_used_at` is how a dead Token is identified. Nested under the owner because a User may only ever manage their own.",
        "responses": {
          "200": {
            "description": "Your Tokens",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Token" } }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Mint a Token",
        "description": "The secret is returned exactly once and stored only as a hash. Tokens do not expire; revoke by deleting. A Token carries no role of its own — its owner's current role is resolved on every request, so a demotion applies to existing Tokens immediately.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "A display name, so a Token can be recognised later."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Minted. `secret` will never be shown again.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenCreated" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" }
        }
      }
    },
    "/users/me/tokens/{token_id}": {
      "parameters": [
        {
          "name": "token_id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" }
        }
      ],
      "delete": {
        "tags": ["users"],
        "summary": "Revoke a Token",
        "description": "The next request bearing it fails.",
        "responses": {
          "204": { "description": "Revoked" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "404": {
            "description": "No such Token belonging to you",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/users/{user_id}": {
      "parameters": [
        {
          "name": "user_id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" }
        }
      ],
      "patch": {
        "tags": ["users"],
        "summary": "Change a User's role",
        "description": "Requires at least the admin role. Refused on the Superadmin's own User (see 409). Takes effect on that User's next request, including requests bearing their existing Tokens.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["role"],
                "properties": { "role": { "$ref": "#/components/schemas/AssignableRole" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/User" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" },
          "404": { "$ref": "#/components/responses/UserNotFound" },
          "409": { "$ref": "#/components/responses/SuperadminProtected" }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Remove a User",
        "description": "Requires at least the admin role. Refused on the Superadmin's own User (see 409). Ends their access and deletes their Tokens. Skills they published remain, and attribution survives through the publisher email snapshot.",
        "responses": {
          "204": { "description": "Removed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" },
          "404": { "$ref": "#/components/responses/UserNotFound" },
          "409": { "$ref": "#/components/responses/SuperadminProtected" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Signed session cookie, http-only, secure, strictly same-site. Carries only the User's id; the role is resolved from the database on every request (ADR-0005). Cannot be revoked before expiry."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A Token, used by the CLI. Looked up by hash on every request; revocable immediately."
      }
    },
...
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
...
    },
    "responses": {
...
      "Unauthenticated": {
        "description": "No valid session or Token. The web interface treats this as an expired session, discards cached state, and returns to login.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    "AdminOnly": {
        "description": "Requires at least the admin role (admin or the superadmin)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "UserNotFound": {
        "description": "No such User",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ValidationFailed": {
        "description": "The request failed a validation rule. `code` names the rule and `field` names what broke it, so the CLI and the web interface can report the same specific reason.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "SuperadminProtected": {
        "description": "Refused: this is the Superadmin. Their role is permanent and their account can never be changed or removed.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Role": {
        "type": "string",
        "enum": ["reader", "writer", "admin", "superadmin"],
        "description": "reader: browse, search, read, retrieve Artifacts. writer: also publish, including replacing any Skill. admin: also delete Skills and manage readers and writers. superadmin: also create and manage Admins. There is exactly one superadmin — set once at `/setup` and never reassigned — so it is never a valid value in a request body; see AssignableRole."
      },
      "AssignableRole": {
        "type": "string",
        "enum": ["reader", "writer", "admin"],
        "description": "The roles a request body may assign to a User. `superadmin` is excluded on purpose: it is set exactly once, by `/setup`, and no route ever grants, changes, or removes it afterwards."
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "email",
          "first_name",
          "last_name",
          "role",
          "must_change_password",
          "created_at"
        ],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "role": { "$ref": "#/components/schemas/Role" },
          "must_change_password": {
            "type": "boolean",
            "description": "True only for a User holding a password that was generated for them and not yet replaced. Meaningful only where a password exists — a User who authenticates through an identity provider has none (ADR-0007)."
          },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Token": {
        "type": "object",
        "required": ["id", "name", "created_at", "last_used_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "last_used_at": { "type": ["string", "null"], "format": "date-time" }
        }
      },
      "TokenCreated": {
        "allOf": [
          { "$ref": "#/components/schemas/Token" },
          {
            "type": "object",
            "required": ["secret"],
            "properties": {
              "secret": {
                "type": "string",
                "description": "Shown once. Never cached by the web interface, and stored only as a hash by the Registry."
              }
            }
          }
        ]
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "A stable machine-readable identifier for what went wrong — for validation failures, the specific rule that failed."
              },
              "message": { "type": "string" },
              "field": {
                "type": "string",
                "description": "The field or file that broke the rule, when there is one."
              }
            }
          }
        }
      }
    }
  }
}